+++ allennlp_hub/pretrained/__init__.py
@@ -1,0 +1,2 @@
+from allennlp_hub.pretrained.allennlp_pretrained import *
+from allennlp_hub.pretrained.allennlp_semparse_pretrained import *
